Board Squares Array Refactor
In order to simplify the logic for moving pieces, I have restructured the board class to have a 2d vector instead of a 1d array. I chose a vector arbitrarily, and this implementation could be done just as easily with arrays in the future. Instead of functions to convert between coordinates and array locations, I have created a map to access the squares of the board in constant time. This allows the squares
 vector to have  double bracket opperator (squares[i][j], which simplifies the operations of moving pieces greatly.

I have also updated the readme.